+++ packages/web/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @workflow/web
 
+## 4.1.11
+
 ## 4.1.10
 
 ### Patch Changes